# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CampusVal — SCU AI Academic Advising API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List SCU courses
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: core
          in: query
          description: Filter to courses fulfilling SCU Core requirements
          schema:
            type: boolean
      responses:
        "200":
          description: Courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
  /courses/{code}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course with prerequisites
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Course detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Not found
...
  /courses/{code}/sections:
    get:
      operationId: listCourseSections
      tags: [courses]
      summary: List scheduled sections (instructors, meeting times, ratings) for a course
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
        - name: term
          in: query
          schema:
            $ref: "#/components/schemas/Term"
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseSection"
...
  /graduation-paths/{type}:
    get:
      operationId: getGraduationPath
      tags: [graduation-paths]
      summary: Get a recommended 3-year or 4-year graduation path template
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [three_year, four_year]
        - name: major
          in: query
          schema:
            type: string
            description: Major code, e.g. "CSE" (Computer Science & Engineering). Defaults to CSE.
      responses:
        "200":
          description: Path template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraduationPath"
...
components:
  schemas:
...
    Term:
      type: string
      enum: [fall, winter, spring, summer]
...
    Course:
      type: object
      properties:
        code:
          type: string
        title:
          type: string
        department:
          type: string
        units:
          type: number
        description:
          type: string
        coreAreas:
          type: array
          items:
            type: string
        offeredTerms:
          type: array
          items:
            $ref: "#/components/schemas/Term"
        difficulty:
          type: string
          enum: [easy, moderate, hard, very_hard]
          nullable: true
        restrictedToColleges:
          type: array
          description: If non-empty, only students in one of these SCU colleges may register
          items:
            type: string
      required:
        - code
        - title
        - department
        - units
        - description
        - coreAreas
        - offeredTerms
        - restrictedToColleges
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            prereqLogic:
              type: string
              description: Human-readable prerequisite expression
            prereqGroups:
              type: array
              description: Groups of OR'd alternatives that are AND'd together
              items:
                type: array
                items:
                  type: string
            corequisites:
              type: array
              items:
                type: string
            notes:
              type: string
              nullable: true
          required:
            - prereqLogic
            - prereqGroups
            - corequisites
...
    ProfessorRating:
      type: object
      description: Aggregated professor data sourced from RateMyProfessors and SCU Schedule Helper
      properties:
        instructor:
          type: string
        overallRating:
          type: number
          description: Average rating 1.0-5.0
          nullable: true
        difficulty:
          type: number
          description: Difficulty 1.0-5.0
          nullable: true
        wouldTakeAgainPercent:
          type: number
          description: Percent of students who would take again (0-100)
          nullable: true
        averageGpa:
          type: number
          description: Historical average GPA in this professor's sections
          nullable: true
        numRatings:
          type: integer
        sourceNote:
          type: string
          description: Provenance, e.g. "RateMyProfessors + SCU Schedule Helper aggregate (cached)"
      required:
        - instructor
        - numRatings
        - sourceNote
    CourseSection:
      type: object
      properties:
        id:
          type: string
          description: e.g. "COEN-12-01-fall-2025"
        courseCode:
          type: string
        sectionNumber:
          type: string
        term:
          $ref: "#/components/schemas/Term"
        year:
          type: integer
        instructor:
          type: string
        meetingDays:
          type: array
          items:
            type: string
            enum: [M, T, W, R, F, S, U]
        startTime:
          type: string
          description: 24h HH:mm
        endTime:
          type: string
          description: 24h HH:mm
        location:
          type: string
        seatsTotal:
          type: integer
        seatsOpen:
          type: integer
        waitlist:
          type: integer
        rating:
          $ref: "#/components/schemas/ProfessorRating"
          nullable: true
      required:
        - id
        - courseCode
        - sectionNumber
        - term
        - year
        - instructor
        - meetingDays
        - startTime
        - endTime
        - location
        - seatsTotal
        - seatsOpen
        - waitlist
...
    GraduationPathQuarter:
      type: object
      properties:
        year:
          type: integer
          description: 1-based year, 1=freshman year
        term:
          $ref: "#/components/schemas/Term"
        label:
          type: string
          description: e.g. "Y1 Fall"
        courses:
          type: array
          items:
            type: string
            description: SCU course code
        plannedUnits:
          type: number
        notes:
          type: string
          nullable: true
      required:
        - year
        - term
        - label
        - courses
        - plannedUnits
    GraduationPath:
      type: object
      properties:
        type:
          type: string
          enum: [three_year, four_year]
        major:
          type: string
        title:
          type: string
        summary:
          type: string
        feasibilityNote:
          type: string
        averageUnitsPerQuarter:
          type: number
        requiresOverload:
          type: boolean
        quarters:
          type: array
          items:
            $ref: "#/components/schemas/GraduationPathQuarter"
        risks:
          type: array
          items:
            type: string
      required:
        - type
        - major
        - title
        - summary
        - feasibilityNote
        - averageUnitsPerQuarter
        - requiresOverload
        - quarters
        - risks
...