# lib/api-zod/src/generated/types/courseSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CampusVal — SCU AI Academic Advising API
 * OpenAPI spec version: 0.1.0
 */
import type { CourseSectionMeetingDaysItem } from "./courseSectionMeetingDaysItem";
import type { ProfessorRating } from "./professorRating";
import type { Term } from "./term";

export interface CourseSection {
  /** e.g. "COEN-12-01-fall-2025" */
  id: string;
  courseCode: string;
  sectionNumber: string;
  term: Term;
  year: number;
  instructor: string;
  meetingDays: CourseSectionMeetingDaysItem[];
  /** 24h HH:mm */
  startTime: string;
  /** 24h HH:mm */
  endTime: string;
  location: string;
  seatsTotal: number;
  seatsOpen: number;
  waitlist: number;
  rating?: ProfessorRating | null;
}